PatternEnhancer/requirements.txt
tensorflow
pillow
tqdm